requirements.txt
biopython
pandas
numpy
streamlit
